refactor: Rename errors file
