+++ zcached/enums.py
